添加功能

+++ src/main/java/com/example/schoolshop/service/impl/CartServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.schoolshop.service.impl;
 
+import cn.dev33.satoken.stp.StpUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
@@ -7,6 +8,7 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.example.schoolshop.common.ErrorCode;
+import com.example.schoolshop.constant.UserConstant;
 import com.example.schoolshop.domain.Cart;
 import com.example.schoolshop.domain.Product;
 import com.example.schoolshop.domain.User;
@@ -38,7 +40,17 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
     private UserService userService;
     @Override
     public Double sum() {
-        List<Cart> carts = this.baseMapper.selectList(null);
+        // 1. 获取session中的信息
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
+        //2.查询数据
+        QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("userId",currentUser.getId());
+        List<Cart> carts = this.baseMapper.selectList(queryWrapper);
         double sum = 0;
         for (Cart cart:carts) {
             sum+=cart.getTotalPrice();
@@ -48,31 +60,40 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
     }
 
     @Override
-    public AddResponse cartAdd(String username, Long product_id,Long num) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public AddResponse cartAdd(Long product_id,Long num) {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
+//        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
+//        userQueryWrapper.eq("username",username);
+//        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
 
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userId", user.getId());
+        queryWrapper.eq("userId", currentUser.getId());
         queryWrapper.eq("productId", product_id);
         if (this.baseMapper.selectOne(queryWrapper) != null) {
             throw new BusinessException(ErrorCode.PARAMS_ERROR, "商品已存在");
         }
 
         Cart cart = new Cart();
-        cart.setUserid(user.getId());
+        cart.setUserid(currentUser.getId());
         cart.setProductId(product_id);
 
         // 查询其他表中的信息
         Product product = productService.getById(product_id);
 
         // 设置其他属性
+        Date createTime = new Date();
         cart.setQuantity(num);
         cart.setProductName(product.getName());
         cart.setUnitPrice(product.getPrice());
         cart.setTotalPrice(product.getPrice()*product.getStock());
         cart.setProductImage(product.getImage());
+        cart.setCreatedTime(createTime);
+        cart.setUpdatedTime(createTime);
         boolean saveResult= this.save(cart);
 
         if (!saveResult) {
@@ -96,37 +117,45 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
     }
 
     @Override
-    public List<Cart> cartQuery(String username) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public List<Cart> cartQuery() {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userId",user.getId());
-
+        queryWrapper.eq("userId",currentUser.getId());
         List<Cart> list = this.baseMapper.selectList(queryWrapper);
         System.out.println(list.toString());
         return list;
     }
 
     @Override
-    public List<Cart> cartPage(String username, Integer start, Integer pageSize) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public List<Cart> cartPage(Integer start, Integer pageSize) {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userId",user.getId());
+        queryWrapper.eq("userId",currentUser.getId());
         IPage<Cart> page = new Page<>(start,pageSize);
         IPage<Cart> resultPage = this.baseMapper.selectPage(page, queryWrapper); // 执行数据库查询操作
         return resultPage.getRecords();
     }
 
     @Override
-    public Boolean cartUpdate(String username, Long productId, Long quantity) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public Boolean cartUpdate( Long productId, Long quantity) {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userid", user.getId());
+        queryWrapper.eq("userid", currentUser.getId());
         queryWrapper.eq("productId", productId);
         Cart cart = this.baseMapper.selectOne(queryWrapper);
 
@@ -135,11 +164,11 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
         }
 
         Double unitPrice = cart.getUnitPrice();
-        Double total = cart.getTotalPrice()*quantity;
+        Double total = unitPrice*quantity;
         Date updateTime = new Date();
 
         UpdateWrapper<Cart> updateWrapper = new UpdateWrapper<>();
-        updateWrapper.eq("userId", user.getId());
+        updateWrapper.eq("userId", currentUser.getId());
         updateWrapper.eq("productId", productId);
         updateWrapper.set("quantity", quantity);
         updateWrapper.set("totalPrice", total);
@@ -150,12 +179,15 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
     }
 
     @Override
-    public Boolean cartDelete(String username, Long product_id) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public Boolean cartDelete(Long product_id) {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userId",user.getId());
+        queryWrapper.eq("userId",currentUser.getId());
         queryWrapper.eq("productId",product_id);
         Cart cart = this.baseMapper.selectOne(queryWrapper);
         if (cart == null) {
@@ -166,18 +198,35 @@ public class CartServiceImpl extends ServiceImpl<CartMapper,Cart> implements Car
     }
 
     @Override
-    public Boolean DeleteAll(String username) {
-        QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        userQueryWrapper.eq("username",username);
-        User user = userService.getBaseMapper().selectOne(userQueryWrapper);
+    public Boolean DeleteAll() {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
         QueryWrapper<Cart> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("userId",user.getId());
+        queryWrapper.eq("userId",currentUser.getId());
         Cart cart = this.baseMapper.selectOne(queryWrapper);
         if (cart == null) {
             throw new BusinessException(ErrorCode.PARAMS_ERROR,"购物车不存在");
         }
         int rows = this.baseMapper.delete(queryWrapper);
         return rows > 0;
+    }
+
+    @Override
+    public List<Cart> QueryAll() {
+        StpUtil.checkLogin();
+        Object userObject = StpUtil.getSession().get(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObject;
+        if (currentUser == null || currentUser.getId() == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN_ERROR);
+        }
+        QueryWrapper<Cart> queryWrapper = new QueryWrapper<>(null);
+        List<Cart> carts = this.baseMapper.selectList(queryWrapper);
+
+        return carts;
     }
